Add basic search for policy arrow map
Inside get_policy_maps we want to be able to find the optimal actions and assign the arrows corresponding to the actions to a cell in a numpy array to be displayed for visulaization. Currently the indeces are found but not yet replaced by the unicode arrows. Not functional status